requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
google-cloud-speech==2.10.0
google-cloud-texttospeech==2.7.0
google-generativeai==0.1.0
openai==0.27.0
python-dotenv==0.19.0
gspread==5.1.1
oauth2client==4.1.3
pdfplumber==0.7.6
numpy==1.21.2
scikit-learn==1.0.2
sentence-transformers==2.2.2
pydantic==1.8.2
